+++ Sources/HIGThemesContract/HIGTheme+Defaults.swift
@@ -34,6 +34,8 @@ extension HIGTheme {
     public var emptyState: any HIGEmptyStateTokens { HIGSystemEmptyStateTokens() }
     public var closeButton: any HIGCloseButtonTokens { HIGSystemCloseButtonTokens() }
     public var modal: any HIGModalTokens { HIGSystemModalTokens() }
+    public var tooltip: any HIGTooltipTokens { HIGSystemTooltipTokens() }
+    public var popover: any HIGPopoverTokens { HIGSystemPopoverTokens() }
     public var progress: any HIGProgressTokens { HIGSystemProgressTokens() }
     public var alert: any HIGAlertTokens { HIGSystemAlertTokens() }
     public var toast: any HIGToastTokens { HIGSystemToastTokens() }